draft

+++ src/transformers/models/glmga/modular_glmga.py
@@ -259,8 +259,7 @@ class GlmgaVideoProcessor(Glm46VVideoProcessor):
             fps (`int` or `float`, *optional*):
                 Target frames to sample per second. Defaults to `self.fps`.
         Returns:
-            np.ndarray:
-                Indices to sample video frames.
+            np.ndarray: Indices to sample video frames.
         """
         if metadata is None or getattr(metadata, "fps", None) is None:
             raise ValueError(
@@ -311,7 +310,7 @@ class GlmgaVideoProcessor(Glm46VVideoProcessor):
         if len(uniq) & 1:
             uniq.append(uniq[-1])
 
-        return np.array(uniq)
+        return np.array(uniq, dtype=int)
 
     def _preprocess(
         self,
